Change wl_os_dupfd_cloexec minfd to be int
The fcntl() argument is defined to be an int and not a long. This does not
matter on most architectures since the value is passed in registers, but
it causes issues on big-endian architectures that pass variadic arguments
on the stack.

+++ src/wayland-os.c
@@ -73,7 +73,7 @@ wl_os_socket_cloexec(int domain, int type, int protocol)
 }
 
 int
-wl_os_dupfd_cloexec(int fd, long minfd)
+wl_os_dupfd_cloexec(int fd, int minfd)
 {
 	int newfd;
